Guava Leaf Disease Detection Web-Application/requirements.txt
Flask==2.3.3
tensorflow==2.18.0
keras==3.8.0
Pillow==11.2.1
numpy==1.26.4
Werkzeug==3.1.3
gunicorn==21.2.0
